new multivalued component
